backend/requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
pyjwt
passlib[bcrypt]